Move SmartArrayInt, delete FileCacheARC and HashQueue.
