+++ grib/src/main/java/ucar/nc2/grib/collection/SmartArrayInt.java
@@ -1,9 +1,9 @@
 /*
- * Copyright (c) 1998-2018 John Caron and University Corporation for Atmospheric Research/Unidata
+ * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
 
-package ucar.nc2.internal.cache;
+package ucar.nc2.grib.collection;
 
 import javax.annotation.concurrent.Immutable;
 import java.util.Arrays;
